# shared/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AmbientMetadata": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "palette": {
            "items": {
              "type": "string"
            },
            "title": "Palette",
            "type": "array"
          },
          "sounds": {
            "items": {
              "type": "string"
            },
            "title": "Sounds",
            "type": "array"
          }
        },
        "required": [
          "description",
          "palette",
          "sounds"
        ],
        "title": "AmbientMetadata",
        "type": "object"
      },
...
      "DialogueResponse": {
        "additionalProperties": false,
        "properties": {
          "character_id": {
            "title": "Character Id",
            "type": "string"
          },
          "character_name": {
            "title": "Character Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "companion",
              "npc"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "character_id",
          "character_name",
          "role",
          "text"
        ],
        "title": "DialogueResponse",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "AUTH_CONFIGURATION_ERROR",
          "INVALID_INPUT",
          "HTTP_ERROR",
          "NOT_FOUND",
          "CONFLICT",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GameIslandState": {
        "additionalProperties": false,
        "properties": {
          "island_id": {
            "const": "joy_meadow",
            "title": "Island Id",
            "type": "string"
          },
          "name": {
            "const": "Joy Meadow",
            "title": "Name",
            "type": "string"
          },
          "restoration_level": {
            "maximum": 100,
            "minimum": 0,
            "title": "Restoration Level",
            "type": "integer"
          },
          "restored": {
            "title": "Restored",
            "type": "boolean"
          }
        },
        "required": [
          "island_id",
          "name",
          "restoration_level",
          "restored"
        ],
        "title": "GameIslandState",
        "type": "object"
      },
...
      "GameStateResponse": {
        "additionalProperties": false,
        "properties": {
          "dialogue": {
            "items": {
              "$ref": "#/components/schemas/DialogueResponse"
            },
            "title": "Dialogue",
            "type": "array"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponse"
            },
            "title": "Inventory",
            "type": "array"
          },
          "island": {
            "$ref": "#/components/schemas/GameIslandState"
          },
          "journal": {
            "items": {
              "$ref": "#/components/schemas/JournalEntryResponse"
            },
            "title": "Journal",
            "type": "array"
          },
          "quest": {
            "$ref": "#/components/schemas/QuestStateResponse"
          },
          "recipe": {
            "$ref": "#/components/schemas/RecipeStateResponse"
          },
          "save": {
            "$ref": "#/components/schemas/SaveStatusResponse"
          },
          "started": {
            "title": "Started",
            "type": "boolean"
          }
        },
        "required": [
          "started",
          "island",
          "dialogue",
          "inventory",
          "quest",
          "recipe",
          "journal",
          "save"
        ],
        "title": "GameStateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryItemResponse": {
        "additionalProperties": false,
        "properties": {
          "collected": {
            "title": "Collected",
            "type": "boolean"
          },
          "ingredient_id": {
            "enum": [
              "vanilla_orchid",
              "honey_bloom"
            ],
            "title": "Ingredient Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quantity": {
            "minimum": 0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "ingredient_id",
          "name",
          "quantity",
          "collected"
        ],
        "title": "InventoryItemResponse",
        "type": "object"
      },
...
      "IslandSummary": {
        "additionalProperties": false,
        "properties": {
          "ambient": {
            "$ref": "#/components/schemas/AmbientMetadata"
          },
          "availability": {
            "enum": [
              "playable",
              "coming_in_version_1"
            ],
            "title": "Availability",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "landmarks": {
            "items": {
              "$ref": "#/components/schemas/LandmarkResponse"
            },
            "title": "Landmarks",
            "type": "array"
          },
          "manifests": {
            "items": {
              "$ref": "#/components/schemas/ManifestReference"
            },
            "title": "Manifests",
            "type": "array"
          },
          "map_order": {
            "minimum": 0,
            "title": "Map Order",
            "type": "integer"
          },
          "map_position": {
            "$ref": "#/components/schemas/WorldMapPosition"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "restoration_level": {
            "maximum": 100,
            "minimum": 0,
            "title": "Restoration Level",
            "type": "integer"
          },
          "restoration_state": {
            "enum": [
              "unrestored",
              "restoring",
              "restored"
            ],
            "title": "Restoration State",
            "type": "string"
          },
          "unlocked": {
            "title": "Unlocked",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "emotion",
          "description",
          "unlocked",
          "availability",
          "restoration_level",
          "restoration_state",
          "map_order",
          "map_position",
          "landmarks",
          "ambient",
          "manifests"
        ],
        "title": "IslandSummary",
        "type": "object"
      },
      "IslandsResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IslandSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "IslandsResponse",
        "type": "object"
      },
...
      "JournalEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "recipe_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Name"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "emotion",
          "recipe_name",
          "created_at"
        ],
        "title": "JournalEntryResponse",
        "type": "object"
      },
...
      "LandmarkResponse": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "description"
        ],
        "title": "LandmarkResponse",
        "type": "object"
      },
...
      "ManifestReference": {
        "additionalProperties": false,
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "manifest_type": {
            "enum": [
              "asset",
              "audio"
            ],
            "title": "Manifest Type",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "key",
          "manifest_type",
          "path"
        ],
        "title": "ManifestReference",
        "type": "object"
      },
...
      "QuestStateResponse": {
        "additionalProperties": false,
        "properties": {
          "can_complete": {
            "title": "Can Complete",
            "type": "boolean"
          },
          "collected_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Collected Ingredients",
            "type": "array"
          },
          "crafted": {
            "title": "Crafted",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "quest_id": {
            "title": "Quest Id",
            "type": "string"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "required_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Required Ingredients",
            "type": "array"
          },
          "status": {
            "enum": [
              "not_started",
              "active",
              "completed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "quest_id",
          "title",
          "description",
          "status",
          "required_ingredients",
          "collected_ingredients",
          "recipe_id",
          "crafted",
          "can_complete"
        ],
        "title": "QuestStateResponse",
        "type": "object"
      },
...
      "RecipeStateResponse": {
        "additionalProperties": false,
        "properties": {
          "ability": {
            "title": "Ability",
            "type": "string"
          },
          "can_craft": {
            "title": "Can Craft",
            "type": "boolean"
          },
          "crafted": {
            "title": "Crafted",
            "type": "boolean"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "lore": {
            "title": "Lore",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "required_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Required Ingredients",
            "type": "array"
          }
        },
        "required": [
          "recipe_id",
          "name",
          "emotion",
          "lore",
          "ability",
          "required_ingredients",
          "crafted",
          "can_craft"
        ],
        "title": "RecipeStateResponse",
        "type": "object"
      },
...
      "SaveStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "last_event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event"
          },
          "last_saved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Saved At"
          },
          "status": {
            "enum": [
              "not_saved",
              "saved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "last_saved_at",
          "last_event"
        ],
        "title": "SaveStatusResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeatherResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WeatherStateResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "WeatherResponse",
        "type": "object"
      },
      "WeatherStateResponse": {
        "additionalProperties": false,
        "properties": {
          "condition": {
            "enum": [
              "sunny",
              "rain",
              "fog",
              "snow",
              "magical"
            ],
            "title": "Condition",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intensity": {
            "maximum": 100,
            "minimum": 0,
            "title": "Intensity",
            "type": "integer"
          },
          "island_id": {
            "format": "uuid",
            "title": "Island Id",
            "type": "string"
          },
          "island_key": {
            "title": "Island Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "island_id",
          "island_key",
          "condition",
          "intensity"
        ],
        "title": "WeatherStateResponse",
        "type": "object"
      },
      "WorldEventResponse": {
        "additionalProperties": false,
        "properties": {
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "island_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Island Key"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "id",
          "event_type",
          "island_key",
          "occurred_at"
        ],
        "title": "WorldEventResponse",
        "type": "object"
      },
      "WorldEventsResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorldEventResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "WorldEventsResponse",
        "type": "object"
      },
      "WorldMapPosition": {
        "additionalProperties": false,
        "properties": {
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "WorldMapPosition",
        "type": "object"
      },
      "WorldResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/WorldEventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "islands": {
            "items": {
              "$ref": "#/components/schemas/IslandSummary"
            },
            "title": "Islands",
            "type": "array"
          },
          "weather": {
            "items": {
              "$ref": "#/components/schemas/WeatherStateResponse"
            },
            "title": "Weather",
            "type": "array"
          }
        },
        "required": [
          "islands",
          "weather",
          "events"
        ],
        "title": "WorldResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Flavor Odyssey Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/world": {
      "get": {
        "operationId": "get_world",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get World",
        "tags": [
          "world"
        ]
      }
    },
    "/api/world/events": {
      "get": {
        "operationId": "get_world_events",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Events",
        "tags": [
          "world"
        ]
      }
    },
    "/api/world/islands": {
      "get": {
        "operationId": "list_world_islands",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IslandsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Islands",
        "tags": [
          "world"
        ]
      }
    },
    "/api/world/islands/{island_id}": {
      "get": {
        "operationId": "get_world_island",
        "parameters": [
          {
            "in": "path",
            "name": "island_id",
            "required": true,
            "schema": {
              "title": "Island Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IslandSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Island",
        "tags": [
          "world"
        ]
      }
    },
    "/api/world/islands/{island_id}/restore": {
      "post": {
        "operationId": "restore_world_island",
        "parameters": [
          {
            "in": "path",
            "name": "island_id",
            "required": true,
            "schema": {
              "title": "Island Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restore Island",
        "tags": [
          "world"
        ]
      }
    },
    "/api/world/weather": {
      "get": {
        "operationId": "get_world_weather",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Weather",
        "tags": [
          "world"
        ]
      }
    },
...
  },
...
}